remove old query engine
