+++ tsdb/engine_test.go
@@ -1,3 +1,3 @@
 package tsdb_test
 
-import _ "github.com/influxdb/influxdb/tsdb/engine"
+// import _ "github.com/influxdb/influxdb/tsdb/engine"
