+++ tsdb/shard_test.go
@@ -1,7 +1,6 @@
 package tsdb_test
 
 import (
-	"fmt"
 	"io/ioutil"
 	"os"
 	"path"
@@ -12,10 +11,11 @@ import (
 
 	"github.com/influxdb/influxdb/models"
 	"github.com/influxdb/influxdb/tsdb"
-	"github.com/influxdb/influxdb/tsdb/engine/b1"
 )
 
 func TestShardWriteAndIndex(t *testing.T) {
+	t.Skip("pending tsm1 iterator impl")
+
 	tmpDir, _ := ioutil.TempDir("", "shard_test")
 	defer os.RemoveAll(tmpDir)
 	tmpShard := path.Join(tmpDir, "shard")
@@ -84,6 +84,8 @@ func TestShardWriteAndIndex(t *testing.T) {
 }
 
 func TestShardWriteAddNewField(t *testing.T) {
+	t.Skip("pending tsm1 iterator impl")
+
 	tmpDir, _ := ioutil.TempDir("", "shard_test")
 	defer os.RemoveAll(tmpDir)
 	tmpShard := path.Join(tmpDir, "shard")
@@ -138,86 +140,6 @@ func TestShardWriteAddNewField(t *testing.T) {
 		t.Fatalf("field names wasn't saved to measurement index")
 	}
 
-}
-
-// Ensure the shard will automatically flush the WAL after a threshold has been reached.
-func TestShard_Autoflush(t *testing.T) {
-	path, _ := ioutil.TempDir("", "shard_test")
-	defer os.RemoveAll(path)
-
-	// Open shard with a really low size threshold, high flush interval.
-	sh := tsdb.NewShard(1, tsdb.NewDatabaseIndex(), filepath.Join(path, "shard"), filepath.Join(path, "wal"), tsdb.EngineOptions{
-		EngineVersion:          b1.Format,
-		MaxWALSize:             1024, // 1KB
-		WALFlushInterval:       1 * time.Hour,
-		WALPartitionFlushDelay: 1 * time.Millisecond,
-	})
-	if err := sh.Open(); err != nil {
-		t.Fatal(err)
-	}
-	defer sh.Close()
-
-	// Write a bunch of points.
-	for i := 0; i < 100; i++ {
-		if err := sh.WritePoints([]models.Point{models.MustNewPoint(
-			fmt.Sprintf("cpu%d", i),
-			map[string]string{"host": "server"},
-			map[string]interface{}{"value": 1.0},
-			time.Unix(1, 2),
-		)}); err != nil {
-			t.Fatal(err)
-		}
-	}
-
-	// Wait for autoflush.
-	time.Sleep(100 * time.Millisecond)
-
-	// Make sure we have series buckets created outside the WAL.
-	if n, err := sh.SeriesCount(); err != nil {
-		t.Fatal(err)
-	} else if n < 10 {
-		t.Fatalf("not enough series, expected at least 10, got %d", n)
-	}
-}
-
-// Ensure the shard will automatically flush the WAL after a threshold has been reached.
-func TestShard_Autoflush_FlushInterval(t *testing.T) {
-	path, _ := ioutil.TempDir("", "shard_test")
-	defer os.RemoveAll(path)
-
-	// Open shard with a high size threshold, small time threshold.
-	sh := tsdb.NewShard(1, tsdb.NewDatabaseIndex(), filepath.Join(path, "shard"), filepath.Join(path, "wal"), tsdb.EngineOptions{
-		EngineVersion:          b1.Format,
-		MaxWALSize:             10 * 1024 * 1024, // 10MB
-		WALFlushInterval:       100 * time.Millisecond,
-		WALPartitionFlushDelay: 1 * time.Millisecond,
-	})
-	if err := sh.Open(); err != nil {
-		t.Fatal(err)
-	}
-	defer sh.Close()
-
-	// Write some points.
-	for i := 0; i < 100; i++ {
-		if err := sh.WritePoints([]models.Point{models.MustNewPoint(
-			fmt.Sprintf("cpu%d", i),
-			map[string]string{"host": "server"},
-			map[string]interface{}{"value": 1.0},
-			time.Unix(1, 2),
-		)}); err != nil {
-			t.Fatal(err)
-		}
-	}
-
-	// Wait for time-based flush.
-	time.Sleep(100 * time.Millisecond)
-
-	// Make sure we have series buckets created outside the WAL.
-	if n, err := sh.SeriesCount(); err != nil {
-		t.Fatal(err)
-	} else if n < 10 {
-		t.Fatalf("not enough series, expected at least 10, got %d", n)
-	}
 }
 
 func BenchmarkWritePoints_NewSeries_1K(b *testing.B)   { benchmarkWritePoints(b, 38, 3, 3, 1) }
