+++ tsdb/show_measurements.go
@@ -1,5 +1,6 @@
 package tsdb
 
+/*
 import (
 	"encoding/json"
 	"fmt"
@@ -245,3 +246,4 @@ func (m *ShowMeasurementsMapper) Close() {
 		m.remote.Close()
 	}
 }
+*/
